refactor(kg): extract cytoscape presentation
